Add jcache tests.
