+++ integrationtests/src/test/resources/log4j2.properties
@@ -71,12 +71,15 @@ logger.quartz.level = WARN
 logger.quartz-job.name = org.axonframework.deadline.quartz.DeadlineJob
 logger.quartz-job.level = OFF
 # Test Containers
-logger.testcontainers.name = org.testcontainers
-logger.testcontainers.level = WARN
-logger.docker-java.name = com.github.dockerjava
-logger.docker-java.level = WARN
-logger.docker-java-http5.name = com.github.dockerjava.zerodep.shaded.org.apache.hc.client5.http.wire
-logger.docker-java-http5.level = OFF
+logger.testcontainers.name=org.testcontainers
+logger.testcontainers.level=WARN
+logger.docker-java.name=com.github.dockerjava
+logger.docker-java.level=WARN
+logger.docker-java-http5.name=com.github.dockerjava.zerodep.shaded.org.apache.hc.client5.http.wire
+logger.docker-java-http5.level=OFF
 # Token store
 logger.in-memory-token-store.name=org.axonframework.eventhandling.tokenstore.inmemory.InMemoryTokenStore
-logger.in-memory-token-store.level = ERROR
+logger.in-memory-token-store.level=ERROR
+# Ehcache
+logger.ehcache.name=org.ehcache
+logger.ehcache.level=WARN
